# Conflicts:
#	.gitignore
#	BUILD.bazel
#	build.sh
#	java/runtime/src/main/java/org/ray/runtime/actor/NativeRayActor.java
#	java/runtime/src/main/java/org/ray/runtime/task/NativeTaskSubmitter.java
#	python/ray/_raylet.pxd
#	python/ray/_raylet.pyx
#	python/ray/experimental/streaming/README.rst
#	python/ray/experimental/streaming/examples/articles.txt
#	python/ray/experimental/streaming/examples/key_selectors.py
#	python/ray/experimental/streaming/examples/simple.py
#	python/ray/experimental/streaming/examples/toy.txt
#	python/ray/experimental/streaming/examples/wordcount.py
#	python/ray/experimental/streaming/operator.py
#	python/ray/experimental/streaming/streaming.py
#	python/ray/ray_constants.py
#	python/ray/streaming
#	python/ray/worker.py
#	src/ray/core_worker/core_worker.cc
#	src/ray/core_worker/core_worker.h
#	src/ray/core_worker/lib/java/org_ray_runtime_actor_NativeRayActor.cc
#	src/ray/core_worker/lib/java/org_ray_runtime_actor_NativeRayActor.h
#	src/ray/core_worker/lib/java/org_ray_runtime_task_NativeTaskSubmitter.cc
#	src/ray/core_worker/lib/java/org_ray_runtime_task_NativeTaskSubmitter.h
#	streaming/README.md
#	streaming/python/README.rst
#	streaming/python/examples/articles.txt
#	streaming/python/examples/key_selectors.py
#	streaming/python/examples/simple.py
#	streaming/python/examples/toy.txt
#	streaming/python/examples/wordcount.py
#	streaming/python/operator.py
#	streaming/python/ray/streaming/README.rst
#	streaming/python/ray/streaming/benchmarks/micro/batched_queue_benchmark.py
#	streaming/python/ray/streaming/communication.py
#	streaming/python/ray/streaming/examples/articles.txt
#	streaming/python/ray/streaming/examples/key_selectors.py
#	streaming/python/ray/streaming/examples/simple.py
#	streaming/python/ray/streaming/examples/toy.txt
#	streaming/python/ray/streaming/examples/wordcount.py
#	streaming/python/ray/streaming/operator.py
#	streaming/python/ray/streaming/streaming.py
#	streaming/python/streaming.py
#	streaming/src/queue/transport.cc
#	streaming/src/queue/transport.h
#	streaming/src/queue/utils.h

+++ src/ray/raylet/object_manager_integration_test.cc
@@ -126,7 +126,7 @@ class TestObjectManagerBase : public ::testing::Test {
 
 class TestObjectManagerIntegration : public TestObjectManagerBase {
  public:
-  uint num_expected_objects;
+  size_t num_expected_objects;
 
   int num_connected_clients = 0;
 
@@ -178,7 +178,7 @@ class TestObjectManagerIntegration : public TestObjectManagerBase {
   void TestPush(int64_t data_size) {
     ray::Status status = ray::Status::OK();
 
-    num_expected_objects = (uint)1;
+    num_expected_objects = (size_t)1;
     ObjectID oid1 = WriteDataToClient(client1, data_size);
     server1->object_manager_.Push(oid1, client_id_2);
   }
